use checked in manifest files
