+++ src/page-modules/widget/available-widgets/GYJwhgtkA/manifest.json
@@ -1,0 +1,22 @@
+{
+  "latest": {
+    "version": "0.2.0",
+    "created": "2023-12-12T20:21:05.475Z",
+    "urls": {
+      "umd": "/widget/GYJwhgtkA/0.2.0/planner-web.umd.js",
+      "esm": "/widget/GYJwhgtkA/0.2.0/planner-web.mjs",
+      "css": "/widget/GYJwhgtkA/0.2.0/planner-web.css"
+    }
+  },
+  "all": [
+    {
+      "version": "0.2.0",
+      "created": "2023-12-12T20:21:05.475Z",
+      "urls": {
+        "umd": "/widget/GYJwhgtkA/0.2.0/planner-web.umd.js",
+        "esm": "/widget/GYJwhgtkA/0.2.0/planner-web.mjs",
+        "css": "/widget/GYJwhgtkA/0.2.0/planner-web.css"
+      }
+    }
+  ]
+}
