+++ src/page-modules/widget/index.ts
@@ -1,2 +1,2 @@
 export { CopyMarkup, CopyMarkupLarge } from './copy-markup';
-export type { PlannerWidgetData, PlannerModule } from './server/data';
+export type { PlannerWidgetData, PlannerModule } from './server';
